Ajout de multer et modif des controllers

+++ controllers/book.js
@@ -6,28 +6,33 @@ const Book = require("../models/Book");
 // sort by model/book order
 
 exports.getAllBooks = (req, res, next) => {
-    Book.find()
-      .then((books) => res.status(200).json(books))
-      .catch((error) => res.status(400).json({ error }));
-  };
+  Book.find()
+    .then((books) => res.status(200).json(books))
+    .catch((error) => res.status(400).json({ error }));
+};
 
 exports.getOneBook = (req, res, next) => {
-    Book.findOne({ _id: req.params.id })
-      .then((book) => res.status(200).json(book))
-      .catch((error) => res.status(404).json({ error }));
-  };
+  Book.findOne({ _id: req.params.id })
+    .then((book) => res.status(200).json(book))
+    .catch((error) => res.status(404).json({ error }));
+};
 
 exports.getBestRating = (req, res, next) => {
-    Book.find()
-      .sort({ averageRating: -1 })
-      .limit(3)
-      .then((books) => res.status(200).json(books))
-      .catch((error) => res.status(400).json({ error }));
-  }
+  Book.find()
+    .sort({ averageRating: -1 })
+    .limit(3)
+    .then((books) => res.status(200).json(books))
+    .catch((error) => res.status(400).json({ error }));
+};
 
 exports.createBook = (req, res, next) => {
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
   const book = new Book({
-    ...req.body,
+    ...bookObject,
+    userId: req.auth.userId,// permet de récupérer l'id de l'utilisateur
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
   });
   book
     .save()
@@ -36,15 +41,34 @@ exports.createBook = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+  const bookObject = req.file
+    ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}}://${req.get("host")}/images/${req.file.filename}`,
+      }
+    : { ...req.body };
+  Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
     .then(() => res.status(200).json({ message: "Objet modifié !" }))
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id})
+       .then(book => {
+           if (book.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = book.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Book.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 exports.addRating = (req, res, next) => {
